Rename prop_types.py -> props.py
